% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiLODmice.R
\name{.multicens.draw3}
\alias{.multicens.draw3}
\title{Base function for imputing multiply left censored data with MICE,}
\usage{
.multicens.draw3(y, ry, x, lod.j, ...)
}
\arguments{
\item{y}{Vector to be imputed}

\item{ry}{Logical vector of length \code{length(y)} indicating the the subset
\code{y[ry]} of elements in y to which the imputation model is fitted.
The \code{ry} generally distinguishes the observed (\code{TRUE}) and
missing values (\code{FALSE}) in y.}

\item{x}{Numeric design matrix with \code{length(y)} rows with predictors
for \code{y}. Matrix \code{x} may have no missing values.}

\item{lod.j}{Vector of varying LODs (left censoring values). -Inf for non censored values.}

\item{...}{Other named arguments.}
}
\value{
parameters
}
\description{
Based on \code{Lapidus et al.} extended to allow multiple LODs. Modified version of .cens.draw3 function from doMIsaul package.
}
\keyword{internal}
