% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendat_multiLOD.R
\name{gendat_multiLOD}
\alias{gendat_multiLOD}
\title{Generate toy data for testing functions.}
\usage{
gendat_multiLOD(
  n = 5000,
  LOD1 = c(-2, -1),
  LOD2 = c(-1, -2),
  rho = 0.5,
  betas = c(1, 1, 1, 1),
  nonlinear = FALSE
)
}
\arguments{
\item{n}{sample size}

\item{LOD1}{vector containing first and second LODs for X1}

\item{LOD2}{vector containing first and second LODs for X2}

\item{rho}{correlation for X's}

\item{betas}{vector containing 5 coefficients}

\item{nonlinear}{should the function use non-linear associations (default=FALSE)}
}
\value{
true data, observed data and LOD data
}
\description{
Function generates toy data with multiple LODs for testing functions.
}
\examples{
set.seed(123)
require(crch)
require(censReg)
require(mvtnorm)

## generate complete data
n <- 5000
dat <- gendat_multiLOD(n)
head(dat$X)
head(dat$df)
head(dat$LODs)

}
