% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiLODmice.R
\name{mice.impute.multicens}
\alias{mice.impute.multicens}
\title{Impute left censored data with MICE}
\usage{
mice.impute.multicens(y, ry, x, lod.j, excludenames, wy = NULL, ...)
}
\arguments{
\item{y}{Vector to be imputed}

\item{ry}{Logical vector of length \code{length(y)} indicating the the subset
\code{y[ry]} of elements in y to which the imputation model is fitted.
The \code{ry} generally distinguishes the observed (\code{TRUE}) and
missing values (\code{FALSE}) in y.}

\item{x}{Numeric design matrix with \code{length(y)} rows with predictors
for \code{y}. Matrix \code{x} may have no missing values.}

\item{lod.j}{Vector of varying LODs (left censoring values). -Inf for non censored values.}

\item{wy}{Logical vector of length \code{length(y)}. A \code{TRUE} value
indicates locations in \code{y} for which imputations are created.}

\item{...}{Other named arguments.}
}
\value{
Vector with imputed data, same type as \code{y}, and of length
\code{sum(wy)}.
}
\description{
Function based on \code{Lapidus et al.} for imputing left-censored data with mice; extended to varying LODs. Modified version of mice.impute.cens function from doMIsaul package.
}
