% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_funs.R
\name{pool_ests}
\alias{pool_ests}
\title{Pool generic estimates.}
\usage{
pool_ests(obj)
}
\arguments{
\item{obj}{list of ests and vcovs to be pooled}
}
\value{
List of estimates vector and variance-covariance matrix.
}
\description{
Internal function for pooling generic estimates via Rubin's rules.
}
\seealso{
\code{\link{pool_lm}}
}
