% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_funs.R
\name{pool_lm}
\alias{pool_lm}
\title{Fit LM/GLM to each imputed dataset and combine.}
\usage{
pool_lm(form, impdat, fam = "gaussian")
}
\arguments{
\item{form}{model formula as in lm/glm}

\item{impdat}{list of imputed datasets}

\item{fam}{family (default="gaussian") to be passed to glm()}
}
\value{
List of estimates vector and variance-covariance matrix.
}
\description{
Uses Rubin's rules to combine LM/GLM fits across imputed datasets.
}
