% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendat_multiLOD.R
\name{gendat_multiLOD}
\alias{gendat_multiLOD}
\title{Generate toy data for testing functions.}
\usage{
gendat_multiLOD(n = 5000)
}
\arguments{
\item{n}{sample size}
}
\value{
true data, observed data and LOD data
}
\description{
Function generates toy data with multiple LODs for testing functions.
}
\examples{
set.seed(123)
require(crch)
require(censReg)
require(mvtnorm)

## generate complete data
n <- 5000
dat <- gendat_multiLOD(n)
head(dat$X)
head(dat$df)
head(dat$LODs)

}
