% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_funs.R
\name{pool_gam}
\alias{pool_gam}
\title{Fit GAM to each imputed dataset and combine.}
\usage{
pool_gam(form, impdat, fam = "gaussian")
}
\arguments{
\item{form}{model formula as in mgcv}

\item{impdat}{list of imputed datasets}

\item{fam}{family (default="gaussian") to be passed to mgcv::gam()}
}
\value{
List of estimates vector and variance-covariance matrix.
}
\description{
Uses Rubin's rules to combine GAM fits across imputed datasets.
}
\examples{
set.seed(123)
require(crch)
require(censReg)
require(mvtnorm)

## generate complete data
n <- 5000
dat <- gendat_multiLOD(n,nonlinear=TRUE)
X <- dat$X
df <- dat$df
LODs <- dat$LODs

## model fit with no censoring
g_gold <- gam(df$y~s(X1)+s(X2)+X3+X4,data=X) ## gold standard
plot(g_gold)

## complete cases only
g_cc <- gam(y~s(X1)+s(X2)+X3+X4,data=df)
plot(g_cc)

## naive imputation: LOD/sqrt(2)
df_naive <- df
df_naive$X1[!is.na(LODs$X1)] <- LODs$X1[!is.na(LODs$X1)]/sqrt(2)
df_naive$X2[!is.na(LODs$X2)] <- LODs$X2[!is.na(LODs$X2)]/sqrt(2)
g_naive <- gam(y~s(X1)+s(X2)+X3+X4,data=df_naive)
plot(g_naive)

## MI
kk <- 20
df_imp <- multiLODmice(data = df, data.lod = LODs,mi.m = kk,excludeLODimp = "y")
g_imp <- pool_gam(y~s(X1,bs="bs",k=10)+s(X2,bs="bs",k=10)+X3+X4,df_imp)
plot(g_imp)

}
\seealso{
\code{\link{pool_ests}} and \code{\link{pool_lm}}
}
