% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_funs.R
\name{pool_gam}
\alias{pool_gam}
\title{Fit GAM to each imputed dataset and combine.}
\usage{
pool_gam(form, impdat, fixedknots, fam = "gaussian", combinemethod = "betas")
}
\arguments{
\item{form}{model formula as in mgcv}

\item{impdat}{list of imputed datasets}

\item{fixedknots}{list of fixed knots to be used by all model fits}

\item{fam}{family (default="gaussian") to be passed to mgcv::gam()}

\item{combinemethod}{method for combining results. Default="betas" combines coefficients as in lm(). Otherwise combines fitted values}
}
\value{
List of estimates vector and variance-covariance matrix.
}
\description{
Uses Rubin's rules to combine GAM fits across imputed datasets.
}
\seealso{
\code{\link{pool_ests}} and \link{pool_lm}
}
