% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_funs.R
\name{pool_lm}
\alias{pool_lm}
\title{Fit LM/GLM to each imputed dataset and combine.}
\usage{
pool_lm(form, impdat, fam = "gaussian")
}
\arguments{
\item{form}{model formula as in lm/glm}

\item{impdat}{list of imputed datasets}

\item{fam}{family (default="gaussian") to be passed to glm()}
}
\value{
List of estimates vector and variance-covariance matrix.
}
\description{
Uses Rubin's rules to combine LM/GLM fits across imputed datasets.
}
\examples{
set.seed(123)
require(crch)
require(censReg)
require(mvtnorm)

## generate complete data
n <- 5000
dat <- gendat_multiLOD(n)
X <- dat$X
df <- dat$df
LODs <- dat$LODs

## model fit with no censoring
g_gold <- lm(df$y~X1+X2+X3+X4,data=X) ## gold standard

## complete cases only
g_cc <- lm(y~X1+X2+X3+X4,data=df)

## naive imputation: LOD/sqrt(2)
df_naive <- df
df_naive$X1[!is.na(LODs$X1)] <- LODs$X1[!is.na(LODs$X1)]/sqrt(2)
df_naive$X2[!is.na(LODs$X2)] <- LODs$X2[!is.na(LODs$X2)]/sqrt(2)
g_naive <- lm(y~X1+X2+X3+X4,data=df_naive)

## MI
kk <- 20
df_imp <- multiLODmice(data = df, data.lod = LODs,mi.m = kk)
g_imp <- pool_lm(y~X1+X2+X3+X4,df_imp)

## compare percent bias
beta <- c(0,1,1,1,1)
round(100*(coef(g_gold)-beta))
round(100*(coef(g_cc)-beta))
round(100*(coef(g_naive)-beta))
round(100*(g_imp$ests-beta))

## compare uncertainty
round(sqrt(diag(vcov(g_gold))),3)
round(sqrt(diag(vcov(g_cc))),3)
round(sqrt(diag(vcov(g_naive))),3)
round(sqrt(diag(g_imp$vcov)),3)


}
\seealso{
\code{\link{pool_ests}} and \code{\link{pool_gam}}
}
